replace address model with state string [#157252772]
