+++ src/ServiceHub.Batch.Library/Models/Batch.cs
@@ -8,7 +8,7 @@ namespace ServiceHub.Batch.Library.Models
     /// <summary>
     /// Batch Model
     /// Contains all information describing any one batch.
-    /// </summary>
+    /// </ssebummary>
     /// <remarks>
     /// A batch refers to a training period for Revature associates. Model includes following information:
     /// unique batch id, start/end date, technology stack, location, number of associates and list of associates.
@@ -16,6 +16,9 @@ namespace ServiceHub.Batch.Library.Models
     [DataContract]
     public class Batch
     {
+        /// <value> State code list to validate state </value>
+        [IgnoreDataMember]
+        private static readonly string[] StateCodes = { "AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DE", "FL", "GA", "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD", "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ", "NM", "NY", "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC", "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY" };
         /// <value> Unique batch id </value>
         [DataMember]
         public Guid BatchId { get; set; }
@@ -34,9 +37,9 @@ namespace ServiceHub.Batch.Library.Models
         /// <value> Batch technology stack </value>
         [DataMember]
         public string BatchSkill { get; set; }
-        /// <value> Address where training takes place </value>
+        /// <value> State where training takes place </value>
         [DataMember]
-        public Address Address { get; set; }
+        public string State { get; set; }
         /// <value> List of associate ids in batch </value>
         [DataMember]
         public List<Guid> UserIds { get; set; }
@@ -56,7 +59,7 @@ namespace ServiceHub.Batch.Library.Models
             BatchName = null;
             BatchOccupancy = null;
             BatchSkill = null;
-            Address = null;
+            State = null;
             UserIds = null;
         }
 
@@ -69,11 +72,12 @@ namespace ServiceHub.Batch.Library.Models
             if (BatchId == Guid.Empty) { return false; }
             if (StartDate == null) { return false; }
             if (EndDate == null) { return false; }
-            if (String.IsNullOrEmpty(BatchName)) { return false; }
+            if (String.IsNullOrEmpty(BatchName) || BatchName.Length > 255) { return false; }
             if (BatchOccupancy == null) { return false; }
-            if (BatchOccupancy < 0 || BatchOccupancy > 100) { return false; }
-            if (String.IsNullOrEmpty(BatchSkill)) { return false; }
-            if (Address == null) { return false; }
+            if (BatchOccupancy < 0) { return false; }
+            if (String.IsNullOrEmpty(BatchSkill) || BatchSkill.Length > 255) { return false; }
+            if (String.IsNullOrEmpty(State)) { return false; }
+            if (!StateCodes.Contains(State.ToUpper())) { return false; }
             if (UserIds?.Any() != true) { return false; }
 
             return true;
